third_party/WebKit/LayoutTests/fast/dom/custom/invalid-type-extension-crash.svg

<html xmlns="http://www.w3.org/1999/xhtml"><x-test is="x-a" id="t"></x-test>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script><![CDATA[
test(() => assert_true(document.getElementById("t") instanceof Element),
     'Accessing an XML element that looks like a Custom Element ' +
     'should not crash.');
]]></script>
</html>
